# Optimize some i64 instructions. 