These instructions are the ones that appear most frequently when verifying ECDSA signatures using Botan.  `i32_wrap_i64` in particular seems to have a big impact. I'm guessing this is because the way it was originally implemented interferes with store/load forwarding.

performance of signature verification before this change:
```
total: 7552749ns, load key: 110086ns, prepare: 926585ns, verify: 6516078ns
total: 8024477ns, load key: 104540ns, prepare: 902859ns, verify: 7017078ns
total: 7398503ns, load key: 105751ns, prepare: 889461ns, verify: 6403291ns
total: 7795229ns, load key: 106230ns, prepare: 894435ns, verify: 6794564ns
total: 7359000ns, load key: 112327ns, prepare: 889713ns, verify: 6356960ns
total: 7654549ns, load key: 90307ns, prepare: 890394ns, verify: 6673848ns
total: 7315964ns, load key: 130398ns, prepare: 935368ns, verify: 6250198ns
total: 7618456ns, load key: 137700ns, prepare: 1023959ns, verify: 6456797ns
total: 7597169ns, load key: 148117ns, prepare: 1012942ns, verify: 6436110ns
total: 7281994ns, load key: 105689ns, prepare: 967185ns, verify: 6209120ns
```

and after:
```
total: 4930000ns, load key: 97293ns, prepare: 734408ns, verify: 4098299ns
total: 4631545ns, load key: 96973ns, prepare: 658758ns, verify: 3875814ns
total: 4679591ns, load key: 79157ns, prepare: 637894ns, verify: 3962540ns
total: 4868950ns, load key: 106273ns, prepare: 722536ns, verify: 4040141ns
total: 4807308ns, load key: 88287ns, prepare: 653077ns, verify: 4065944ns
total: 4638457ns, load key: 97420ns, prepare: 655816ns, verify: 3885221ns
total: 4618231ns, load key: 77771ns, prepare: 648436ns, verify: 3892024ns
total: 4955404ns, load key: 79110ns, prepare: 644343ns, verify: 4231951ns
total: 4733457ns, load key: 138950ns, prepare: 703320ns, verify: 3891187ns
total: 4571457ns, load key: 87475ns, prepare: 653438ns, verify: 3830544ns
```